Restriciones

+++ src/main/java/org/springframework/samples/petclinic/web/DaycareController.java
@@ -1,8 +1,14 @@
 package org.springframework.samples.petclinic.web;
 
+import java.time.LocalDate;
+import java.time.ZoneId;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.GregorianCalendar;
 import java.util.Map;
 
 import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.Daycare;
 import org.springframework.samples.petclinic.model.Pet;
@@ -15,7 +21,6 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.InitBinder;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
@@ -36,11 +41,20 @@ public class DaycareController {
 	@InitBinder
 	public void setAllowedFields(WebDataBinder dataBinder) {
 		dataBinder.setDisallowedFields("id");
+		dataBinder.setDisallowedFields("aforo");
 	}
 	
 	@GetMapping(value= "/owners/{ownerId}/pets/{petId}/deleteDaycare/{daycareId}")
 	public String deleteDaycare(@PathVariable("ownerId") int ownerId, @PathVariable("daycareId") final int daycareId, final Pet pet, final ModelMap model) {
-		this.daycareService.delete(daycareId);
+		Daycare daycare= this.daycareService.findDaycareById(daycareId);
+		Calendar calendar = new GregorianCalendar();
+		calendar.add(Calendar.DAY_OF_MONTH, 2);
+		Date minimumDeadline = calendar.getTime();
+		if(daycare.getDate().isBefore(minimumDeadline.toInstant().atZone(ZoneId.systemDefault()).toLocalDate())){
+			return "redirect:/owners/" + ownerId;
+		} else {
+			this.daycareService.delete(daycareId);
+		}
 		return "redirect:/owners/" + ownerId;
 	}
   
@@ -61,9 +75,20 @@ public class DaycareController {
 			return "daycares/createOrUpdateDaycareForm";
 		}
 		else {
-			this.petService.saveDaycare(daycare);	
-			return "redirect:/owners/{ownerId}";
+			if(this.daycareService.oneDaycareById(daycare.getDate(), petId)==1){
+                result.rejectValue("date", "", "This daycare already exists");
+    			return "daycares/createOrUpdateDaycareForm";
+			} else if(this.daycareService.countDaycareByDate(daycare.getDate())==daycare.getCapacity()) {
+				result.rejectValue("date", "", "Complete capacity. Please, select another date");
+    			return "daycares/createOrUpdateDaycareForm";
+			
+			} else if(daycare.getDate().isBefore(LocalDate.now())) {
+				result.rejectValue("date", "", "Please, select a future date");
+    			return "daycares/createOrUpdateDaycareForm";
+			}
 		}
+		this.petService.saveDaycare(daycare);	
+		return "redirect:/owners/{ownerId}";
 	}
 	
 	@GetMapping(value = "/owners/{ownerId}/pets/{petId}/daycares/{daycareId}/edit")
